refactor: put date of birth validation functions to "lib" folder
